Switched from Toasts to ListView for message display and a few other small items.
Messages are now displayed below the button area instead of as System Toasts.
Removed an *.iml that doesn't need to be checked in.
Updated the .gitignore to prevent build time files from accidentally being able to be checked in.
Switched to dark Holo style since it looks a little nicer.

Reviewed in https://codereview.appspot.com/12969044/